requirements.txt
aiohttp==3.7.4.post0
anyio==3.4.0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
discord==1.7.3
discord.py==1.7.3
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
multidict==5.2.0
pafy==0.5.5
pornhub-api==0.2.0
pycparser==2.21
pydantic==1.8.2
PyNaCl==1.4.0
python-dotenv==0.19.2
python-twitch-client==0.7.1
pytube==11.0.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2
youtube-dl==2021.6.6
youtube-search-python==1.5.2
